# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job-App Backend API",
    "description": "REST API for job search platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/me/profile": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Profile",
        "description": "Retrieve the user's saved profile.\n\nEndpoint: GET /me/profile\nAuth: Required (JWT)\n\nResponse (HTTP 200):\n    - perfilEstructurado: Structured profile\n    - resumenParaMatching: Summary text for matching (or null)\n    - cargosSugeridos: Suggested roles (or empty list)\n    - cargosActivos: User-selected active roles (or empty list)\n    - profileVersion: Version number\n    - updatedAt: Timestamp of last update\n    - resumenGenerating: Boolean flag (read-only from DB)\n\nError Responses:\n    - HTTP 404: Profile not found\n\nLogic:\n1. Query Perfiles table by userId (pk)\n2. If not found: Return HTTP 404 with profile_not_found\n3. Return all stored fields (including resumenGenerationStatus → resumenGenerating)\n4. Log retrieval status\n\nRequirements:\n- 2.1, 2.2: Query Perfiles by userId\n- 2.3: Return perfilEstructurado\n- 2.4: Return resumenParaMatching (may be null)\n- 2.5: Return cargosActivos, cargosSugeridos, profileVersion, updatedAt\n- 3.5: Return resumenGenerating (read-only boolean, NOT persisted by this endpoint)\n- 3.7: Return 404 with profile_not_found if missing",
        "operationId": "get_profile_me_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Profile Me Profile Get"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Save Profile",
        "description": "Save a structured profile for the user.\n\nEndpoint: PUT /me/profile\nAuth: Required (JWT)\n\nRequest Body:\n    - perfilEstructurado: Structured profile (PerfilEstructurado model)\n\nResponse (HTTP 200):\n    - profileVersion: Updated version number\n    - updatedAt: Timestamp of update\n    - cargosActivos: Returned as-is (not modified by this endpoint)\n\nError Responses:\n    - HTTP 400: Validation error\n\nLogic:\n1. Validate request body (already done by Pydantic)\n2. Get current profile to determine next version\n3. Persist ONLY perfilEstructurado, profileVersion (incremented), updatedAt\n4. Return HTTP 200 with version info immediately (no async background tasks)\n5. Never modify resumenParaMatching, resumenGenerationStatus, cargosSugeridos\n6. Log save operation\n\nRequirements:\n- 2.1: Validate body against SaveProfileRequest\n- 2.3: Persist perfilEstructurado\n- 2.4: Increment profileVersion\n- 2.5: Update updatedAt timestamp\n- 3.1: Return HTTP 200 with version/timestamp\n- 3.2: Perform write atomically, return immediately\n- 3.3: Log profile save operation\n- 10.1: Perform DynamoDB write\n- 10.2: Increment version atomically\n- 10.3: Never enqueue background tasks in this endpoint\n- 10.4: Never modify resumenParaMatching or resumenGenerationStatus\n- 10.5: Return only version and updatedAt (worker owns other fields)",
        "operationId": "save_profile_me_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Save Profile Me Profile Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Certificacion": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre",
            "description": "Certification name"
          },
          "emisor": {
            "type": "string",
            "title": "Emisor",
            "description": "Issuing organization"
          },
          "ano": {
            "type": "string",
            "title": "Ano",
            "description": "Year obtained"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "emisor",
          "ano"
        ],
        "title": "Certificacion",
        "description": "Certification entry from CV."
      },
...
      "Educacion": {
        "properties": {
          "titulo": {
            "type": "string",
            "title": "Titulo",
            "description": "Degree or certification title"
          },
          "institucion": {
            "type": "string",
            "title": "Institucion",
            "description": "Institution name"
          },
          "ano": {
            "type": "string",
            "title": "Ano",
            "description": "Year or year range"
          },
          "especializacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Especializacion",
            "description": "Specialization or focus area"
          }
        },
        "type": "object",
        "required": [
          "titulo",
          "institucion",
          "ano"
        ],
        "title": "Educacion",
        "description": "Education entry from CV."
      },
      "ExperienciaLaboral": {
        "properties": {
          "puesto": {
            "type": "string",
            "title": "Puesto",
            "description": "Job title"
          },
          "empresa": {
            "type": "string",
            "title": "Empresa",
            "description": "Company name"
          },
          "duracion": {
            "type": "string",
            "title": "Duracion",
            "description": "Duration (e.g., '2 years', 'Jan 2020 - Dec 2021')"
          },
          "descripcion": {
            "type": "string",
            "title": "Descripcion",
            "description": "Job description and responsibilities"
          },
          "tecnologias": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tecnologias",
            "description": "Technologies used"
          }
        },
        "type": "object",
        "required": [
          "puesto",
          "empresa",
          "duracion",
          "descripcion"
        ],
        "title": "ExperienciaLaboral",
        "description": "Work experience entry from CV."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PerfilEstructurado": {
        "properties": {
          "experiencia": {
            "items": {
              "$ref": "#/components/schemas/ExperienciaLaboral"
            },
            "type": "array",
            "title": "Experiencia",
            "description": "Work experience entries"
          },
          "educacion": {
            "items": {
              "$ref": "#/components/schemas/Educacion"
            },
            "type": "array",
            "title": "Educacion",
            "description": "Education entries"
          },
          "proyectos": {
            "items": {
              "$ref": "#/components/schemas/Proyecto"
            },
            "type": "array",
            "title": "Proyectos",
            "description": "Project entries"
          },
          "certificaciones": {
            "items": {
              "$ref": "#/components/schemas/Certificacion"
            },
            "type": "array",
            "title": "Certificaciones",
            "description": "Certification entries"
          },
          "skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills",
            "description": "Technical skills"
          },
          "lenguajes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lenguajes",
            "description": "Languages spoken/read"
          }
        },
        "type": "object",
        "required": [
          "skills"
        ],
        "title": "PerfilEstructurado",
        "description": "User's structured CV profile, parsed from raw CV text.\n\nThis is the primary data structure for storing user's professional information."
      },
      "Proyecto": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre",
            "description": "Project name"
          },
          "descripcion": {
            "type": "string",
            "title": "Descripcion",
            "description": "Project description"
          },
          "tecnologias": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tecnologias",
            "description": "Technologies used"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url",
            "description": "Project URL (optional)"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "descripcion"
        ],
        "title": "Proyecto",
        "description": "Project entry from CV."
      },
      "SaveProfileRequest": {
        "properties": {
          "perfilEstructurado": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PerfilEstructurado"
              }
            ],
            "description": "Structured profile object from CV parsing or manual editing"
          }
        },
        "type": "object",
        "required": [
          "perfilEstructurado"
        ],
        "title": "SaveProfileRequest",
        "description": "Request body for PUT /me/profile.\n\nContains the structured profile to be saved."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}